+++ CHANGELOG.md
@@ -9,6 +9,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Allow defining callback-driven variable nodes with `Server::add_data_source_variable_node()`.
 - Allow deletion of server nodes with `Server::delete_node()`.
 - Add `BADINTERNALERROR` and `BADWRITENOTSUPPORTED` variants to `ua::StatusCode`.